requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyJWT==2.8.0
boto3>=1.34.0
psycopg2-binary==2.9.9